requirements.txt
uvicorn[standard]==0.24.0
alembic==1.13.0
pydantic==2.7.4
python-dotenv==1.0.0
asyncpg==0.29.0

langchain==0.3.0
langgraph==0.4.8
langgraph-prebuilt==0.2.3
langchain-anthropic>=0.1.16
langchain_community
pypdf
langchain-openai